+++ include/cask/observable/FlatMapObserver.hpp
@@ -74,7 +74,7 @@ Task<Ack,None> FlatMapObserver<TI,TO,E>::onNext(const TI& value) {
 
 template <class TI, class TO, class E>
 Task<None,None> FlatMapObserver<TI,TO,E>::onError(const E& error) {
-    if(!completed.test_and_set()) {
+    if(!completed.test_and_set(std::memory_order_relaxed)) {
         return downstream->onError(error);
     } else {
         return Task<None,None>::none();
@@ -83,7 +83,7 @@ Task<None,None> FlatMapObserver<TI,TO,E>::onError(const E& error) {
 
 template <class TI, class TO, class E>
 Task<None,None> FlatMapObserver<TI,TO,E>::onComplete() {
-    if(!completed.test_and_set()) {
+    if(!completed.test_and_set(std::memory_order_relaxed)) {
         return downstream->onComplete();
     } else {
         return Task<None,None>::none();
@@ -92,7 +92,7 @@ Task<None,None> FlatMapObserver<TI,TO,E>::onComplete() {
 
 template <class TI, class TO, class E>
 Task<None,None> FlatMapObserver<TI,TO,E>::onCancel() {
-    if(!completed.test_and_set()) {
+    if(!completed.test_and_set(std::memory_order_relaxed)) {
         return downstream->onCancel();
     } else {
         return Task<None,None>::none();