+++ include/cask/observable/MapTaskObserver.hpp
@@ -56,7 +56,7 @@ Task<Ack,None> MapTaskObserver<TI,TO,E>::onNext(const TI& value) {
 
 template <class TI, class TO, class E>
 Task<None,None> MapTaskObserver<TI,TO,E>::onError(const E& error) {
-    if(!completed.test_and_set()) {
+    if(!completed.test_and_set(std::memory_order_relaxed)) {
         return downstream->onError(error);
     } else {
         return Task<None,None>::none();
@@ -65,7 +65,7 @@ Task<None,None> MapTaskObserver<TI,TO,E>::onError(const E& error) {
 
 template <class TI, class TO, class E>
 Task<None,None> MapTaskObserver<TI,TO,E>::onComplete() {
-    if(!completed.test_and_set()) {
+    if(!completed.test_and_set(std::memory_order_relaxed)) {
         return downstream->onComplete();
     } else {
         return Task<None,None>::none();
@@ -74,7 +74,7 @@ Task<None,None> MapTaskObserver<TI,TO,E>::onComplete() {
 
 template <class TI, class TO, class E>
 Task<None,None> MapTaskObserver<TI,TO,E>::onCancel() {
-    if(!completed.test_and_set()) {
+    if(!completed.test_and_set(std::memory_order_relaxed)) {
         return downstream->onCancel();
     } else {
         return Task<None,None>::none();