+++ src/cask/Erased.cpp
@@ -1,0 +1,3 @@
+#include "cask/Erased.hpp"
+
+cask::Pool<cask::erased_pool_block_size> cask::Erased::pool(cask::erased_pool_startup_blocks);